% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVclass.R
\docType{class}
\name{CNVvault-class}
\alias{CNVvault-class}
\alias{CNVvault}
\title{CNVvault Class}
\arguments{
\item{Segments}{Dataframe of your segments.}

\item{SampleInfo}{Information on your clusters}

\item{Plot}{the plot of everything}
}
\description{
Process data for copynumber plot
}
\details{
This class will hold your segments and sample information
}
\examples{
CNVvault()
}
\keyword{CNV}
\keyword{Copynumber}
