% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTestData.R
\name{generateTestData}
\alias{generateTestData}
\title{generateTestData}
\usage{
generateTestData(genome = "hg19", FinalChrom = "chrX")
}
\arguments{
\item{genome}{string. The geneome to download from UCSC. Default: hg19}

\item{FinalChrom}{string. The final chromosome to include in the plot. Default: chrX}
}
\description{
Generate some basic test data
}
\details{
This script will generate some very basic test data to check thr script is running correctly
}
\examples{
generateTestData()

}
\keyword{CNV}
\keyword{Copynumber}
