% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenomeLengths.R
\name{getGenomeLengths}
\alias{getGenomeLengths}
\title{setPositionsCNV}
\usage{
getGenomeLengths(ReturnClass, genome = "hg19", FinalChrom = "chrX")
}
\arguments{
\item{genome}{string. The geneome to download from UCSC. Default: hg19}

\item{FinalChrom}{string. The final chromosome to include in the plot. Default: chrX}

\item{Copynumber}{Dataframe of your segments. REQUIRED}
}
\description{
Set the start and coords for the geom_rect
}
\details{
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will output a full plot of all the segments
}
\examples{
setPositionsCNV(CNVvault_class)
}
\keyword{CNV}
\keyword{Copynumber}
