% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderCNV.R
\name{orderCNV}
\alias{orderCNV}
\title{orderCNV}
\usage{
orderCNV(ReturnClass, descending = TRUE)
}
\arguments{
\item{ReturnClass}{Class of type CNVvault created using prepareCNV. REQUIRED}

\item{descending}{sets the direction of the ordering. Default: TRUE}
}
\description{
Process data for copynumber plot
}
\details{
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will create the sample info dataframe and set ordering for the samples
}
\examples{
orderCNV(CNVvault_object)
}
\keyword{CNV}
\keyword{Copynumber}
