% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCopynumber.R
\name{plotCopynumber}
\alias{plotCopynumber}
\title{plotCopynumber}
\usage{
plotCopynumber(ReturnClass, setColours = c(), ...)
}
\arguments{
\item{ReturnClass}{Class of type CNV. REQUIRED}

\item{setColours}{A vector of colours for the plot}
}
\description{
Print Genome wide copynumber plot
}
\details{
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will output a full plot of all the segments
}
\examples{
Plot_Copynumber(CNVvault_Object)
}
\keyword{CNV}
\keyword{Copynumber}
