% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCNV.R
\name{prepareCNV}
\alias{prepareCNV}
\title{prepareCNV}
\usage{
prepareCNV(
  Copynumber,
  sampleID = "sampleID",
  genome = "hg19",
  chr = "chr",
  start.pos = "start.pos",
  end.pos = "end.pos",
  calls = "calls"
)
}
\arguments{
\item{Copynumber}{Dataframe of your segments. REQUIRED}

\item{sampleID}{string: name of the column for the sample IDs. Default: "sampleID"}

\item{genome}{The genome to use for plotting. Currently only works on hg19. Default: "hg19"}

\item{chr}{string: name of the column for the chromosome. Default: "chr"}

\item{start.pos}{string: name of the column for the start positions of the segments. Default: "start.pos"}

\item{end.pos}{string: name of the column for the end positions of the segments. Default: "end.pos"}

\item{calls}{string: name of the column for the SV type of the segments. Default: "calls"}
}
\description{
Process data for copynumber plot
}
\details{
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will create an instance of the CNVvault class
}
\examples{
prepareCNV()


}
\keyword{CNV}
\keyword{Copynumber}
