% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPositionsCNV.R
\name{setPositionsCNV}
\alias{setPositionsCNV}
\title{setPositionsCNV}
\usage{
setPositionsCNV(ReturnClass, genome = "hg19")
}
\arguments{
\item{genome}{The genome to use for plotting. Currently only works on hg19. Default: "hg19"}

\item{Copynumber}{Dataframe of your segments. REQUIRED}
}
\description{
Set the start and coords for the geom_rect
}
\details{
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will output a full plot of all the segments
}
\examples{
setPositionsCNV(CNVvault_class)
}
\keyword{CNV}
\keyword{Copynumber}
