% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTestData.R
\name{generateTestData}
\alias{generateTestData}
\title{generateTestData}
\description{
Generate some basic test data
}
\details{
This script will generate some very basic test data to check thr script is running correctly
}
\examples{
generateTestData()

}
\keyword{CNV}
\keyword{Copynumber}
