% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderCNV.R
\name{orderCNV}
\alias{orderCNV}
\title{orderCNV}
\usage{
orderCNV(ReturnClass, order = TRUE, descending = TRUE, sampleOrder = NULL, ...)
}
\arguments{
\item{ReturnClass}{Class of type CNVvault created using prepareCNV. REQUIRED}

\item{order}{Should the samples be ordered or not}

\item{descending}{sets the direction of the ordering. Default: TRUE}

\item{sampleOrder}{(optional) vector of sample names. Order samples should be plotted}
}
\description{
Process data for copynumber plot
}
\details{
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will create the sample info dataframe and set ordering for the samples
}
\examples{
orderCNV(CNVvault_object)
}
\keyword{CNV}
\keyword{Copynumber}
