% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCNV.R
\name{prepareCNV}
\alias{prepareCNV}
\title{prepareCNV}
\usage{
prepareCNV(
  Copynumber,
  sampleID = "sampleID",
  genome = "hg19",
  chr = "chr",
  start.pos = "start.pos",
  end.pos = "end.pos",
  calls = "calls",
  minimum_width = 0,
  ...
)
}
\arguments{
\item{Copynumber}{Dataframe of your segments. REQUIRED}

\item{sampleID}{string: name of the column for the sample IDs. Default: "sampleID"}

\item{genome}{string. The geneome to download from UCSC. Default: hg19}

\item{chr}{string: name of the column for the chromosome. Default: "chr"}

\item{start.pos}{string: name of the column for the start positions of the segments. Default: "start.pos"}

\item{end.pos}{string: name of the column for the end positions of the segments. Default: "end.pos"}

\item{calls}{string: name of the column for the SV type of the segments. Default: "calls"}

\item{minimum_width}{numeric: minimum width for segments, segments shorter than this will be extended to this length. Default: 0 (off)}

\item{FinalChrom}{string. The final chromosome to include in the plot. Default: chrX}
}
\description{
Process data for copynumber plot
}
\details{
This function will take a dataframe with sample, chr, start.pos, end.pos and call.
It will create an instance of the CNVvault class
}
\examples{
prepareCNV()


}
\keyword{CNV}
\keyword{Copynumber}
